+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+2023.6.0 - 30-05-2023
+===================
 ### Features
 * Core | Add `toRenko()` tests [#124](https://github.com/Reiryoku-Technologies/Mida/pull/124)
 * Core | Expose `MidaPeriodUtilities` and add `toRenko` function for converting periods to renko [#123](https://github.com/Reiryoku-Technologies/Mida/pull/123)